Made file structure flatter, and less java-like. No real benefit in only having one class per file.. Just makes searching harder.
