+++ symboard/file_writers.py
@@ -24,3 +24,4 @@ class KeylayoutFileWriter(FileWriter):
         _file.close()
 
         pass
+
