+++ symboard/parsers.py
@@ -4,6 +4,11 @@
 own, but instead to be inherited from and extended from.
 '''
 
+# Third party imports
+from yaml import safe_load
+from typing import Dict
+
+
 class Parser:
     @staticmethod
     def parse(content):
@@ -20,3 +25,10 @@ class FileParser(Parser):
     @staticmethod
     def parse(file_path: str):
         pass
+
+
+class YamlFileParser(FileParser):
+    @staticmethod
+    def parse(file_path: str) -> Dict:
+        with open(file_path, 'r') as stream:
+            return safe_load(stream)
